Correctly print array and object SQL parameters in case of an exception.
Laravel tries to show SQL that caused an exception. While doing it
it also prints all SQL parameters. If there are arrays or objects
in parameter it throws an "Object of class stdClass
could not be converted to string" exception.

This commit json_encodes all non scalar parameters so that they can
be printed in exception.